Update list template fiels by template name tests
